Event View and EventList listeners
You can now touch an event on the event list go to a simple view of the
event.

I couldn't figure out how to implement the listener outside the
EventListAdapter class, so it's not in the TouchManager class like all the
other OnClickListeners. If somebody else figures out how to do it feel
free to move the code.