New Activities - Major Refactor
Did a major refactor. Now instead of just using one activity and switching
its contentView, in most cases we create new activities when creating new
screens. This will allow users to use the back button on their phone as
well as any other Android API calls that navigate the activity stack. It
was a huge pain in the ass.